+++ hypervisor/arch/x86/apic.c
@@ -112,7 +112,7 @@ static u32 apic_ext_features(void)
 		return 0;
 }
 
-int phys_processor_id(void)
+unsigned long phys_processor_id(void)
 {
 	return apic_ops.read_id();
 }
@@ -121,7 +121,7 @@ int apic_cpu_init(struct per_cpu *cpu_data)
 {
 	unsigned int xlc = MAX((apic_ext_features() >> 16) & 0xff,
 			       APIC_REG_XLVT3 - APIC_REG_XLVT0 + 1);
-	unsigned int apic_id = phys_processor_id();
+	unsigned int apic_id = (unsigned int)phys_processor_id();
 	unsigned int cpu_id = cpu_data->cpu_id;
 	unsigned int n;
 	u32 ldr;
