+++ hypervisor/printk.c
@@ -18,7 +18,7 @@
 #include <asm/spinlock.h>
 
 volatile unsigned long panic_in_progress;
-unsigned int panic_cpu = -1;
+unsigned long panic_cpu = -1;
 
 static DEFINE_SPINLOCK(printk_lock);
 
@@ -40,7 +40,7 @@ void printk(const char *fmt, ...)
 
 void panic_printk(const char *fmt, ...)
 {
-	unsigned int cpu_id = phys_processor_id();
+	unsigned long cpu_id = phys_processor_id();
 	va_list ap;
 
 	if (test_and_set_bit(0, &panic_in_progress) && panic_cpu != cpu_id)
